Transportes

+++ src/EmpresaTransportes.java
@@ -2,7 +2,7 @@ import java.util.ArrayList;
 
 public class EmpresaTransportes {
     private String local;
-    private float raioDeAcao;
+    private double raioDeAcao;
     private ArrayList<Transportes> transportes;
     private ArrayList<Encomenda> registos;
     private int numeroMinimoEncomenda;
