+++ src/Encomenda.java
@@ -1,7 +1,7 @@
 import java.util.ArrayList;
 
 public class Encomenda {
-    private float peso;
+    private double peso;
     private String comprador;
     private String vendedor;
     private ArrayList<String> produtos;
